Allow to restore as much as possible in case if only partial transaction logs are present and user enforces recovery.

+++ community/wal/src/main/java/org/neo4j/kernel/impl/transaction/log/entry/LogHeader.java
@@ -56,7 +56,7 @@ public class LogHeader
     @Override
     public int hashCode()
     {
-        int result = (int) logFormatVersion;
+        int result = logFormatVersion;
         result = 31 * result + (int) (logVersion ^ (logVersion >>> 32));
         result = 31 * result + (int) (lastCommittedTxId ^ (lastCommittedTxId >>> 32));
         return result;
